@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>CodeMooc.Web</RootNamespace>
     <AssemblyName>CodeMooc.Web</AssemblyName>
   </PropertyGroup>
@@ -12,12 +12,9 @@
 
   <ItemGroup>
     <PackageReference Include="BCrypt.Net-Next.StrongName" Version="3.2.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.TagHelpers" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-    <PackageReference Include="MySql.Data" Version="8.0.18" />
-    <PackageReference Include="MySql.Data.EntityFrameworkCore" Version="8.0.18" />
-    <PackageReference Include="RestSharp" Version="106.6.10" />
+    <PackageReference Include="MySql.Data" Version="8.0.19" />
+    <PackageReference Include="MySql.Data.EntityFrameworkCore" Version="8.0.19" />
+    <PackageReference Include="RestSharp" Version="106.10.2-alpha.0.8" />
     <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-beta0007" />
   </ItemGroup>
   
@@ -39,12 +36,6 @@
     </Compile>
   </ItemGroup>
   
-  <ItemGroup>
-    <Content Update="Views\Static\Front.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-  </ItemGroup>
-
   <ItemGroup>
     <EmbeddedResource Update="Resources\BadgeEvidenceDescriptions.resx">
       <Generator>PublicResXFileCodeGenerator</Generator>
