@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc.ViewFeatures;
-using Newtonsoft.Json;
+﻿using System.Text.Json;
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
 
 namespace CodeMooc.Web {
 
@@ -8,13 +8,15 @@ namespace CodeMooc.Web {
         public static void Put<T>(this ITempDataDictionary tempData, T value)
             where T : class {
             var key = typeof(T).Name;
-            tempData[key] = JsonConvert.SerializeObject(value);
+            tempData[key] = JsonSerializer.Serialize<T>(value);
         }
 
         public static T Get<T>(this ITempDataDictionary tempData) where T : class {
             var key = typeof(T).Name;
             tempData.TryGetValue(key, out object o);
-            return o == null ? null : JsonConvert.DeserializeObject<T>((string)o);
+
+            string value = o as string;
+            return value == null ? null : JsonSerializer.Deserialize<T>(value);
         }
 
     }
