@@ -1,1 +1,4 @@
-﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+﻿@using CodeMooc.Web.Controllers
+@using CodeMooc.Web.Data
+@using CodeMooc.Web.Model
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
